Task 4 - Retro Game Fixed compiler warnings. Added credits fade.

+++ Task4/Task4/RetroGame/src/States/MainMenu/GamesListMenu.cpp
@@ -104,8 +104,8 @@ void GamesListMenuState::Draw()
 			int y = GetScreenHeight()/2 + (i*32);
 			if (j)
 			{
-				x += sinf(GetTime() * 8) * 4;
-				y += -cosf(GetTime() * 8) * 4;
+				x += (int)(sinf((float)GetTime() * 8) * 4);
+				y += (int)(-cosf((float)GetTime() * 8) * 4);
 			}
 			DrawText(optionText.str().c_str(), (int)m_menuOffset + x, y, fontSize, WHITE);
 		}
