+++ Task4/Task4/RetroGame/src/States/MainMenu/OptionsMenu.cpp
@@ -225,8 +225,8 @@ void OptionsMenuState::Draw()
 			if (j)
 			{
 				// Wobble when highlighted.
-				x += sinf(GetTime() * 8) * 4;
-				y += -cosf(GetTime() * 8) * 4;
+				x += (int)(sinf((float)GetTime() * 8) * 4);
+				y += (int)(-cosf((float)GetTime() * 8) * 4);
 			}
 			DrawText(optionText.str().c_str(), (int)m_menuOffset + x, y, fontSize, WHITE);
 		}
